refactor: Enum으로 request matcher 작성
